ajustes finales

+++ app/Http/Controllers/RoleController.php
@@ -14,6 +14,7 @@ class RoleController extends Controller
      */
     public function index()
     {
+        $this->authorize("viewAny_role");
         $roles = Role::with('permisos')->get();
 
         return response()->json($roles, 200);
@@ -27,6 +28,7 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize("create_role");
         $request->validate([
             "nombre" => "required|unique:roles"
         ]);
@@ -47,6 +49,7 @@ class RoleController extends Controller
      */
     public function show($id)
     {
+        $this->authorize("view_role");
         $rol = Role::with('permisos')->find($id);
         return response()->json($rol, 201);
     }
@@ -60,6 +63,7 @@ class RoleController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $this->authorize("update_role");
         $request->validate([
             "nombre" => "required|unique:roles,nombre,$id"
         ]);
@@ -80,6 +84,7 @@ class RoleController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize("delete_role");
         $rol = Role::find($id);
         $rol->delete();
